requirements.txt
django
django-channels
djangorestframework
channels_redis
twilo
-e git+https://github.com/MeNsaaH/transitionMatrix
